third_party/blink/web_tests/flag-specific/highdpi/accessibility/scroll-to-global-point-nested-expected.txt

Tests that scrolling an element to a specific point can successfully scroll multiple nested scrolling views'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

5000-pixel box
5000-pixel box
5000-pixel box
Target
5000-pixel box
5000-pixel box
5000-pixel box
PASS window.pageYOffset is 0
PASS outerContainer.scrollTop is 0
PASS innerContainer.scrollTop is 0
PASS target.getBoundingClientRect().top is >= 15000
PASS target.getBoundingClientRect().top is 0
FAIL target.getBoundingClientRect().top should be 300. Was 200.
FAIL target.getBoundingClientRect().top should be 3000. Was 2000.
PASS successfullyParsed is true

TEST COMPLETE
